all: add support for MSVC 2005

+++ inc/load.h
@@ -15,7 +15,6 @@
 #pragma once
 
 #include <haiyajan.h>
-#include <stdint.h>
 
 /**
  * Loads a file for the libretro core.
@@ -24,7 +23,7 @@
  * \param ctx	Libretro core context.
  * \return	0 on success, else failure. Use SDL_GetError().
  */
-uint_fast8_t load_libretro_file(struct core_ctx_s *restrict ctx);
+int load_libretro_file(struct core_ctx_s *restrict ctx);
 
 /**
  * Loads a libretro core and assigns its functions to the given libretro
@@ -34,7 +33,7 @@ uint_fast8_t load_libretro_file(struct core_ctx_s *restrict ctx);
  * \param ctx		Libretro core context to initialise.
  * \return		0 on success, else failure. Use SDL_GetError().
  */
-uint_fast8_t load_libretro_core(const char *restrict so_file,
+int load_libretro_core(const char *restrict so_file,
 	struct core_ctx_s *restrict ctx);
 
 /**
