+++ inc/play.h
@@ -45,7 +45,7 @@ void play_frame(struct core_ctx_s *ctx);
  * \param ctx	Libretro core context.
  * \returns	0 on success, else failure. Use SDL_GetError().
  */
-uint_fast8_t play_init_av(struct core_ctx_s *ctx, SDL_Renderer *rend);
+int play_init_av(struct core_ctx_s *ctx, SDL_Renderer *rend);
 
 /**
  * Free audio and video contexts for libretro core.
